move fieldInfoSection component. extract treeTableStyles
