+++ frontend/src/metabase/transforms/pages/TransformInspectPage/components/LensSections/GenericSummarySections/styles.ts
@@ -1,0 +1,5 @@
+export const treeTableStyles = {
+  headerRow: {
+    backgroundColor: "var(--mb-color-background-secondary)",
+  },
+};
