+++ frontend/src/metabase/transforms/pages/TransformInspectPage/components/index.ts
@@ -1,3 +1,2 @@
-export { FieldInfoSection } from "./FieldInfoSection/FieldInfoSection";
 export { LensContent } from "./LensContent/LensContent";
 export { LensNavigation } from "./LensNavigation/LensNavigation";
